requirements.txt
Flask
Flask-Cors
instaloader
scikit-learn
pandas
joblib
google-generativeai
python-dotenv
